Moved from routes/stocks to utils
